Move geography.py to repo layer
