+++ app/api/api_v1/routers/geographies.py
@@ -4,8 +4,8 @@ from fastapi import APIRouter, Depends, HTTPException, status
 
 from app.api.api_v1.schemas.geography import GeographyStatsDTO
 from app.clients.db.session import get_db
-from app.db.crud.geography import get_world_map_stats
 from app.errors import RepositoryError
+from app.repository.geography import get_world_map_stats
 
 _LOGGER = logging.getLogger(__file__)
