+++ app/db/crud/document.py
@@ -34,7 +34,7 @@ from app.api.api_v1.schemas.document import (
 )
 from app.core.lookups import doc_type_from_family_document_metadata
 from app.core.util import to_cdn_url
-from app.db.crud.geography import get_geo_subquery
+from app.repository.geography import get_geo_subquery
 
 _LOGGER = logging.getLogger(__file__)
